chore: rename specialBlockquotesPlugin to messageBlocksPlugin
